requirements.txt
Django
django-opensearch-dsl
djangorestframework
opensearch
opensearch-dsl
psycopg2
haystack
django-haystack
django-cors-headers
pandas